PlayerManager script implemented, to handle prefabs which was Instantiated by using the UtilityPrefab which didn't let Unity to build the project since it is a UnityEditor method.
Fixed some other little issues, so right now the current state of the game is running on Android and it is "playable". So the chest is controled by Motion tracking. Walls are spawning randomly, if  the Player gets hit by the wall it falls into the pool behind the character.