Seems like been able to fix the lagg Issue for now.
I took off (delted a.k.a commented out the mUnityPlayer.onPause() line from the Overridden onpause method of the UnityPlayerActivity class.
Don't know exactly right now why this helped, but i am happy with it and I will try to figure it out later. Especially because I am not sure if it has or has no potentially bad harm on other parts of the application.
The movement directions of some particular bodyparts still seems wrong.